+++ parasol_cpu/benches/cardio.rs
@@ -49,7 +49,7 @@ fn generate_args(enc: &Encryption, sk: &SecretKey) -> CallData<UInt<8, L1GlweCip
         .sum::<u8>();
 
     ArgsBuilder::new()
-        .arg(UInt::<8, _>::encrypt_secret(flags as u64, enc, sk))
+        .arg(UInt::<8, _>::encrypt_secret(flags as u128, enc, sk))
         .arg(UInt::<8, _>::encrypt_secret(40, enc, sk))
         .arg(UInt::<8, _>::encrypt_secret(50, enc, sk))
         .arg(UInt::<8, _>::encrypt_secret(70, enc, sk))
